Refactor: useModel 로직 분할
