+++ src/2_pages/diary/ui/diaryInfo/model/useOpenModal.js
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+export const useOpenModal = () => {
+    
+    const [ isOpenModal, setIsOpenModal ] = useState( false );
+    const onClickModal = () => setIsOpenModal( !isOpenModal );
+
+    return { isOpenModal, onClickModal };
+}
